update myInfo axios
